Changed program name
