--- src/main/java/duke/command/TodoCommand.java
@@ -1,52 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.Task;
-import duke.task.TaskList;
-import duke.task.Todo;
-import duke.ui.Ui;
-
-import java.io.IOException;
-
-/**
- * Handless actions taken when user
- * creates a to-do task.
- */
-public class TodoCommand extends Command {
-    private final Todo todo;
-
-    /**
-     * Instantiates the to-do instance.
-     *
-     * @param todo
-     */
-    public TodoCommand(Todo todo) {
-        this.todo = todo;
-    }
-
-    /**
-     * Adds the to-do task to task list, updates the storage,
-     * then printing to the command line the added task.
-     *
-     * @param taskList Current task list instance.
-     * @param ui Current ui instance.
-     * @param storage Current storage instance.
-     * @throws IOException An error with writing to storage.
-     */
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) throws IOException {
-            taskList.addTasks(todo);
-            storage.updateStorage(taskList);
-            ui.printAddTask(taskList, todo);
-    }
-
-    /**
-     * Adding to-do task does not terminate program.
-     *
-     * @return False.
-     */
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
