--- src/main/java/duke/exception/DukeException.java
@@ -1,12 +1,0 @@
-package duke.exception;
-
-/**
- * Main class to handle the exceptions thrown by Duke.
- */
-public class DukeException extends Exception {
-
-    public DukeException(String message) {
-        super(message);
-    }
-
-}
