--- src/main/java/duke/exception/EmptyDescException.java
@@ -1,12 +1,0 @@
-package duke.exception;
-
-/**
- * An exception when a valid user command does not
- * contain a description.
- */
-public class EmptyDescException extends DukeException {
-    String taskType;
-    public EmptyDescException(String message) {
-        super(message);
-    }
-}
