--- src/main/java/duke/exception/EmptyStorageException.java
@@ -1,12 +1,0 @@
-package duke.exception;
-
-/**
- * Command when there isn't an existing data.
- */
-public class EmptyStorageException extends DukeException {
-
-    public EmptyStorageException (String message) {
-        super(message);
-    }
-
-}
