--- src/main/java/duke/exception/UnknownCommandException.java
@@ -1,10 +1,0 @@
-package duke.exception;
-
-/**
- * An exception when a command isn't valid.
- */
-public class UnknownCommandException extends DukeException {
-    public UnknownCommandException(String message) {
-        super(message);
-    }
-}
