+++ src/main/java/lele/exception/LeleException.java
@@ -1,0 +1,12 @@
+package lele.exception;
+
+/**
+ * Main class to handle the exceptions thrown by Duke.
+ */
+public class LeleException extends Exception {
+
+    public LeleException(String message) {
+        super(message);
+    }
+
+}
