+++ src/main/java/lele/exception/UnknownCommandException.java
@@ -1,0 +1,10 @@
+package lele.exception;
+
+/**
+ * An exception when a command isn't valid.
+ */
+public class UnknownCommandException extends LeleException {
+    public UnknownCommandException(String message) {
+        super(message);
+    }
+}
