--- src/test/java/duke/DukeTest.java
@@ -1,4 +1,0 @@
-package duke;
-
-public class DukeTest {
-}
