+++ src/test/java/lele/LeleTest.java
@@ -1,0 +1,4 @@
+package lele;
+
+public class LeleTest {
+}
